+++ Genius-core/genius2.0/bin/genius.cmd
@@ -1,0 +1,3 @@
+java -DMETA_PATH=../property/meta.data -Dfile.encoding=UTF-8 ^
+-cp ../lib/boot-helper-1.0.jar ^
+name.yalsooni.boothelper.Run %*
