--- Genius-core/src/main/java/name/yalsooni/genius/core/boothelper/classloader/exception/JARFileNotFoundException.java
@@ -1,11 +1,0 @@
-package name.yalsooni.genius.core.boothelper.classloader.exception;
-
-/**
- * Created by ijyoon on 2015. 12. 16..
- */
-public class JARFileNotFoundException extends Exception{
-
-    public JARFileNotFoundException(String dirPath){
-        super("JAR file not found in '"+dirPath+"'");
-    }
-}
