--- Genius-core/src/main/java/name/yalsooni/genius/core/boothelper/classloader/exception/NotDirectoryException.java
@@ -1,11 +1,0 @@
-package name.yalsooni.genius.core.boothelper.classloader.exception;
-
-/**
- * Created by ijyoon on 2015. 12. 16..
- */
-public class NotDirectoryException extends Exception{
-
-    public NotDirectoryException(String dirPath){
-        super("'" + dirPath + "' is not directory.");
-    }
-}
