--- Genius-core/src/main/java/name/yalsooni/genius/core/boothelper/definition/Code.java
@@ -1,13 +1,0 @@
-package name.yalsooni.genius.core.boothelper.definition;
-
-/**
- * 부트헬퍼 에러코드 정의
- * Created by ijyoon on 2016. 7. 29..
- */
-public interface Code {
-    String BH_0001 = "[BH-0001] 메타데이터 파일 로드 기타 예외 발생.";
-    String BH_0002 = "[BH-0002] 메타데이터 파일을 찾을 수 없습니다.";
-    String BH_0003 = "[BH-0003] 프로퍼티 파일을 읽는 중 예외 발생.";
-    String BH_0004 = "[BH-0004] 클래스 로드 중 예외 발생.";
-    String BH_0005 = "[BH-0005] 클래스 호출 진행 중 예외 발생.";
-}
