--- Genius-core/src/main/java/name/yalsooni/genius/core/boothelper/definition/PropertyName.java
@@ -1,15 +1,0 @@
-package name.yalsooni.genius.core.boothelper.definition;
-
-/**
- * Created by ijyoon on 2016. 7. 29..
- */
-public interface PropertyName {
-    String METAFILE_NAME = "meta.data";
-
-    String META = "META";
-    String MAIN_CLASS_NAME = "CLASSNAME.EXE.MAIN";
-    String PROPERTY_ROOT_PATH = "PATH.PROPERTY.ROOT";
-    String LIB_ROOT_PATH = "PATH.LIB.ROOT";
-
-    String PROPERTY_EXT_NAME = ".data";
-}
