--- Genius-core/src/main/java/name/yalsooni/genius/core/common/definition/LifeCycle.java
@@ -1,6 +1,0 @@
-package name.yalsooni.genius.core.common.definition;
-
-public interface LifeCycle {
-    void initialize() throws Exception;
-    void execute() throws Exception;
-}
