+++ Genius-core/src/main/java/name/yalsooni/genius/runner/definition/ErrCode.java
@@ -1,0 +1,14 @@
+package name.yalsooni.genius.runner.definition;
+
+/**
+ * 지니어스 에러코드 정의
+ * @author ijyoon
+ *
+ */
+public interface ErrCode {
+
+	String G_003_0001 = "[G-002-0001] 초기화 오류 입니다.";
+	String G_003_0002 = "[G-002-0002] 실행 오류 입니다.";
+	String G_003_0006 = "[G-002-0006] 파라미터 값의 데이터 타입이 유효하지 않습니다.";
+
+}
