+++ Genius-core/src/main/java/name/yalsooni/genius/runner/definition/LifeCycle.java
@@ -1,0 +1,6 @@
+package name.yalsooni.genius.runner.definition;
+
+public interface LifeCycle {
+    void initialize() throws Exception;
+    void execute() throws Exception;
+}
