+++ Genius-core/src/main/resources/meta.data
@@ -1,0 +1,4 @@
+# Meta Data Set
+CLASSNAME.EXE.MAIN      = name.yalsooni.genius.runner.execute.GeniusExecutor
+PATH.PROPERTY.ROOT      = ../property
+PATH.LIB.ROOT           = ../lib
